[BE] rename PlayerDao and otherDaos
